+++ app/Example.hs
@@ -11,5 +11,11 @@ import Data.List
 
 $(loadEnv)
 
+-- We could also have been a little more explicit about it.
+-- load ["sleep", "echo", "cat"]
+
 main :: IO ()
-main = (sleep 1 >> echo "Hello" >> sleep 2) |> cat
+main = do
+    -- Crash the program if we are missing any executables.
+    [] <- missingExecutables
+    (sleep 1 >> echo "Hello" >> sleep 2) |> cat