+++ shell.nix
@@ -10,7 +10,7 @@ let
       }:
       mkDerivation {
         pname = "shh";
-        version = "0.1.0.0";
+        version = "0.1.1.0";
         src = ./.;
         isLibrary = true;
         isExecutable = true;