+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Revision history for execpath
 
+## 0.1.1.0 -- 2019-01-22
+
+* Generate an IO action to check for missing dependencies at runtime.
+
+* Add ability to use TH to load a list of specific executables.
+
 ## 0.1.0.1 -- 2019-01-22
 
 * Force output of lazy read functions to normal form to prevent